+++ cxx_mkspecs/emscripten_common.py
@@ -19,11 +19,8 @@ def mkspec_emscripten_configure(conf, major, minor, minimum=False,
     # The path to the emscripten compiler
     paths = conf.get_tool_option('emscripten_path')
 
-    # The node.js binary is called "nodejs" on Linux
-    if "linux" in sys.platform:
-        conf.find_program('nodejs', var='NODEJS')
-    else:
-        conf.find_program('node', var='NODEJS')
+    # The node.js binary can be "nodejs" or simply "node"
+    conf.find_program(['nodejs', 'node'], var='NODEJS')
 
     # Find the clang++ compiler
     cxx = conf.find_program(['em++'], path_list=paths)
@@ -87,7 +84,7 @@ def mkspec_emscripten_configure(conf, major, minor, minimum=False,
 
 
 @conf
-def check_emscripten_version(conf, emscripten_cc, major, minor, minium):
+def check_emscripten_version(conf, emscripten_cc, major, minor, minimum):
     try:
         p = subprocess.Popen(
             emscripten_cc + ['--version'],
@@ -98,11 +95,13 @@ def check_emscripten_version(conf, emscripten_cc, major, minor, minium):
     except:
         conf.fatal('could not determine the compiler version')
 
-    if minium and cc_major != major or cc_minor != minor:
-        conf.fatal("Wrong version number: major={0} and minor={1}, "
-                   "expected major={2} and minor={3}."
-                   .format(cc_major, cc_minor, major, minor))
-    elif cc_major < major or (cc_major == major and cc_minor < minor):
-        conf.fatal("Compiler version: major={1} and minor={2}, "
-                   "required minimum: major={1} and minor={2}."
-                   .format(cc_major, cc_minor, major, minor))
+    if minimum:
+        if cc_major < major or (cc_major == major and cc_minor < minor):
+            conf.fatal("Compiler version: major={1} and minor={2}, "
+                       "required minimum: major={1} and minor={2}."
+                       .format(cc_major, cc_minor, major, minor))
+    else:
+        if cc_major != major or cc_minor != minor:
+            conf.fatal("Wrong version number: major={0} and minor={1}, "
+                       "expected major={2} and minor={3}."
+                       .format(cc_major, cc_minor, major, minor))